Add day 13 part 2 solution (and lots of optimization)
That was tricky. Now the code spends most of the time in
DictionaryIterator.next(), so storing the data in a container with
constant lookup will make it way faster again.
Even now I'm down from >10 minutes (didn't leave it running for longer)
to less than a second for a release build.